Move the `SupportedOSPlatformAttribute` to a distinct namespace
