+++ Source/Testably.Abstractions.AccessControl/FileStreamAclExtensions.cs
@@ -1,5 +1,4 @@
 ﻿using System.IO;
-using System.Runtime.Versioning;
 using System.Security.AccessControl;
 using Testably.Abstractions.FileSystem;
