+++ Source/Testably.Abstractions.Interface/FileSystem/IFileInfo.cs
@@ -1,5 +1,4 @@
 ﻿using System.IO;
-using System.Runtime.Versioning;
 
 namespace Testably.Abstractions.FileSystem;
