+++ Source/Testably.Abstractions.Testing/Storage/InMemoryStorage.cs
@@ -7,7 +7,6 @@ using System.Linq;
 using Testably.Abstractions.FileSystem;
 using Testably.Abstractions.Testing.FileSystem;
 using Testably.Abstractions.Testing.Internal;
-using static Testably.Abstractions.Testing.MockFileSystem;
 
 namespace Testably.Abstractions.Testing.Storage;
