+++ Source/Testably.Abstractions/FileSystem/DriveInfoWrapper.cs
@@ -1,6 +1,5 @@
 ﻿using System.Diagnostics.CodeAnalysis;
 using System.IO;
-using System.Runtime.Versioning;
 
 namespace Testably.Abstractions.FileSystem;
