+++ Source/Testably.Abstractions/Usings.cs
@@ -1,0 +1,5 @@
+﻿#if NETSTANDARD2_0 || NETSTANDARD2_1
+global using Testably.Abstractions.Polyfills;
+#else
+global using System.Runtime.Versioning;
+#endif
