+++ Tests/Testably.Abstractions.Tests/TestHelpers/Usings.cs
@@ -4,3 +4,8 @@ global using System;
 global using Testably.Abstractions.Testing;
 global using Testably.Abstractions.Tests.TestHelpers;
 global using Xunit;
+#if NET472
+global using Testably.Abstractions.Polyfills;
+#else
+global using System.Runtime.Versioning;
+#endif
